newVersion

+++ app/UseCases/CommentService.php
@@ -4,40 +4,37 @@ namespace App\UseCases;
 
 use App\Models\Comment;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 
 class CommentService
 {
-    public function create(Request $request)
+    public function create($data)
     {
-        $request->validate([
-           'application_id'=>'required|integer|exists:applications,id',
-           'description'=>'required|string',
-           'column_id'=>'required|string',
-           'version'=>'required|string',
-           'documents'=>'nullable|array|exists:files,id'
-        ]);
         
-        $comment = Comment::make($request->only('name', 'manufacturer', 'model', 'version', 'documents'));
-        $comment->save();
+        
+        $comment = Comment::create($data);
+        
         return $comment;
     }
 
     public function edit(Request $request, Comment $comment)
     {
         $request->validate([
-            'name'=>'required|string',
-            'manufacturer'=>'required|string',
-            'model'=>'required|string',
-            'version'=>'required|string',
-            'documents'=>'nullable|array|exists:files,id'
+            'application_id'=>'required|integer|exists:applications,id',
+            'description'=>'required|string',
+            'column_id'=>'required|string',
         ]);
-        $comment->update($request->only('name', 'manufacturer', 'model', 'version', 'documents'));
+        if(Auth::user()->id==$comment->author){
+            $comment->update($request->only('name', 'manufacturer', 'model', 'version', 'documents'));
+         }
         return $comment;
 
     }
     public function remove(Comment $comment)
     {
-        $comment->delete();
+        if(Auth::user()->id==$comment->author){
+            $comment->delete();
+         }
         return 'deleted';
     }
